% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EleChemr.R
\name{CottrCheck}
\alias{CottrCheck}
\title{Cottrel current check for the Chronoamperometric simulation}
\usage{
CottrCheck(Elefun)
}
\arguments{
\item{Elefun}{the function to be checked = ChronAmp, PotStep}
}
\value{
A graph G/Gcot vs t for the simulation data selected
}
\description{
Return a graph G/Gcot vs t of the electrochemical process
}
\examples{
CottrCheck(ChronAmp(errCheck = TRUE, Method = "BI"))

}
