% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative.R
\name{invMat}
\alias{invMat}
\title{Inverse matrix}
\usage{
invMat(A)
}
\arguments{
\item{A}{matrix to be inverted}
}
\value{
inverse matrix of the selected
}
\description{
Returns the inverse matrix of the selected one
}
\examples{
invMat(A = matrix(c(1,2,6,14), nrow = 2))

}
