% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative.R
\name{OneMat}
\alias{OneMat}
\title{Starting Matrix of oxidazed species}
\usage{
OneMat(i, j = i)
}
\arguments{
\item{i}{number of rows}

\item{j}{number of columns}
}
\value{
a matrix of dimention ixj filled with 1 value
}
\description{
Return a matrix ixj filled with 1 value
}
\examples{
OneMat(2,2)

}
