% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative.R
\name{ZeroMat}
\alias{ZeroMat}
\title{Starting Matrix of reduces species and fluxes}
\usage{
ZeroMat(i, j = i)
}
\arguments{
\item{i}{number of rows}

\item{j}{number of columns}
}
\value{
a matrix of dimention ixj filled with 1 value
}
\description{
Return a matrix ixj filled with 0 value
}
\examples{
ZeroMat(2,2)

}
